update mstest run scripts for CI with new test assembly resolution algoithm

+++ Dev/TestScripts/Tests/MSTest/CI/Run-Activity-Designers-Unit-Tests.ps1
@@ -1,8 +1,4 @@
-﻿if ([string]::IsNullOrEmpty($PSScriptRoot)) {
-	$PSScriptRoot = Split-Path $MyInvocation.MyCommand.Path -Parent
-}
-$SolutionDir = (Get-Item $PSScriptRoot).parent.parent.parent.parent.FullName
-# Read playlists and args.
+﻿# Read playlists and args.
 $TestList = ""
 if ($Args.Count -gt 0) {
     $TestList = $Args.ForEach({ "," + $_ })
@@ -25,8 +21,50 @@ if ($Args.Count -gt 0) {
     }
 }
 
-# Create full VSTest argument string.
-$FullArgsList = " `"$WorkspaceDir\Dev2.Activities.Designers.Tests.dll`" /logger:trx" + $TestList
+# Find test assemblies
+if (Test-Path "$PSScriptRoot\Dev2.Activities.Designers.Tests\bin\Debug\Dev2.Activities.Designers.Tests.dll") {
+	$TestAssemblyPath = "$PSScriptRoot\Warewolf.Tests\Dev2.Activities.Designers.Tests.dll"
+}
+if (Test-Path "$PSScriptRoot\..\Dev2.Activities.Designers.Tests\bin\Debug\Dev2.Activities.Designers.Tests.dll") {
+	$TestAssemblyPath = "$PSScriptRoot\..\Warewolf.Tests\Dev2.Activities.Designers.Tests.dll"
+}
+if (Test-Path "$PSScriptRoot\..\..\Dev2.Activities.Designers.Tests\bin\Debug\Dev2.Activities.Designers.Tests.dll") {
+	$TestAssemblyPath = "$PSScriptRoot\..\..\Warewolf.Tests\Dev2.Activities.Designers.Tests.dll"
+}
+if (Test-Path "$PSScriptRoot\..\..\..\Dev2.Activities.Designers.Tests\bin\Debug\Dev2.Activities.Designers.Tests.dll") {
+	$TestAssemblyPath = "$PSScriptRoot\..\..\..\Warewolf.Tests\Dev2.Activities.Designers.Tests.dll"
+}
+if (Test-Path "$PSScriptRoot\..\..\..\..\Dev2.Activities.Designers.Tests\bin\Debug\Dev2.Activities.Designers.Tests.dll") {
+	$TestAssemblyPath = "$PSScriptRoot\..\..\..\..\Warewolf.Tests\Dev2.Activities.Designers.Tests.dll"
+}
+if (Test-Path "$PSScriptRoot\Dev2.Activities.Designers.Tests.dll") {
+	$TestAssemblyPath = "$PSScriptRoot\Dev2.Activities.Designers.Tests.dll"
+}
+if (Test-Path "$PSScriptRoot\..\Dev2.Activities.Designers.Tests.dll") {
+	$TestAssemblyPath = "$PSScriptRoot\..\Dev2.Activities.Designers.Tests.dll"
+}
+if (Test-Path "$PSScriptRoot\..\..\Dev2.Activities.Designers.Tests.dll") {
+	$TestAssemblyPath = "$PSScriptRoot\..\..\Dev2.Activities.Designers.Tests.dll"
+}
+if (Test-Path "$PSScriptRoot\..\..\..\Dev2.Activities.Designers.Tests.dll") {
+	$TestAssemblyPath = "$PSScriptRoot\..\..\..\Dev2.Activities.Designers.Tests.dll"
+}
+if (Test-Path "$PSScriptRoot\..\..\..\..\Dev2.Activities.Designers.Tests.dll") {
+	$TestAssemblyPath = "$PSScriptRoot\..\..\..\..\Dev2.Activities.Designers.Tests.dll"
+}
+if (Test-Path "$PSScriptRoot\..\..\..\..\..\Dev2.Activities.Designers.Tests.dll") {
+	$TestAssemblyPath = "$PSScriptRoot\..\..\..\..\..\Dev2.Activities.Designers.Tests.dll"
+}
+if (!(Test-Path $TestAssemblyPath)) {
+	Write-Host Cannot find Dev2.Activities.Designers.Tests.dll at $PSScriptRoot\Dev2.Activities.Designers.Tests\bin\Debug or $PSScriptRoot
+	exit 1
+}
+if (!(Test-Path $PSScriptRoot\TestResults)) {
+	New-Item -ItemType Directory $PSScriptRoot\TestResults
+}
+
+# Create full MSTest argument string.
+$FullArgsList = " /testcontainer:`"" + $TestAssemblyPath + "`" /resultsfile:`"" + $PSScriptRoot + "\TestResults\ActivityDesignersUnitTestResults.trx`"" + $TestList
 
 # Write full command including full argument string.
 Out-File -LiteralPath $PSScriptRoot\RunTests.bat -Encoding default -InputObject `"$env:vs140comntools..\IDE\MSTest.exe`"$FullArgsList
